+++ documentation/changes.md
@@ -1,2 +1,10 @@
-- 相比原来的添加了jwt，需要在postman的header里添加```token```项才能用
+- 需要后台挂一个MySQL，并设置好表（不过没有也能启动项目，一读数据库就报错）
+  - Need a MySQL in the background  
+
+
+- 相比原来的添加了jwt token，需要在postman的header里添加```token```项才能用
   - Compared with the original version, jwt token has been added. You need to add the ```token``` item to the postman request header to use it.
+
++ 需要后台挂一个Redis（没有就不能启动项目）
+  + Need a Redis in the background
+